refactor: consolidate date helpers into utils/dateUtils
