+++ src/utils/dateUtils.ts
@@ -1,3 +1,14 @@
+export type DateFormat =
+  | "YYYY-MM-DD"
+  | "YYYY.MM.DD"
+  | "DD/MM/YYYY"
+  | "MM/DD/YYYY"
+  | "DD-MM-YYYY"
+  | "MM-DD-YYYY"
+  | "YYYY/MM/DD"
+  | "DD.MM.YYYY"
+  | "MM.DD.YYYY";
+
 type OperationFormat =
   | "same"
   | "greater"
@@ -129,4 +140,17 @@ const compareYears = (
         compareYears(date1, date2, "same") || compareYears(date1, date2, "less")
       );
   }
+};
+
+export const convertToDateString = (date: Date, format: DateFormat): string => {
+  const day = String(date.getDate()).padStart(2, "0");
+  const month = String(date.getMonth() + 1).padStart(2, "0");
+  const year = date.getFullYear().toString();
+
+  const newDate = format
+    .replace("DD", day)
+    .replace("MM", month)
+    .replace("YYYY", year);
+
+  return newDate;
 };
